nearby_chat_backend/requirements.txt
Django>=4.2
djangorestframework
firebase-admin
python-dotenv
geopy
celery
redis
django-cors-headers
